src/backend/requirements.txt
# 智搭后端服务 - Python依赖

flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
colorthief==0.2.1
gunicorn==21.2.0
# zhipuai==2.0.1  # 已注释，使用Gemini替代
google-generativeai==0.8.3
